Move SQS driver to new namespace
